# performance: track large tx execution cost